Netstatus: Average out throughput when time elapsed is greater than 1 second.
This isn't very accurate but seems to be what other tools that check at slower rates than every second do.  For instance if you are checking every 10 seconds and in that time frame 500 kb was received, was the throughput for each second 50 kb? No, not likely, but that is the result you will see here.  I guess this is one of those close enough things.  Thanks to @davemds for reporting this issue.