shot - move save of shot and upload to slave binary tool and tidy code
this gets rid of that annoying freeze in e due to the sync save and
possibly sync dns lookups in curl. slave binary handles it now and
talks back to e via stdout.

this also splits up the code into more files for specific purposes to
make it easier to add features to and maintain. given on the todo is
the ability to explicitly select a crop region and do some basic
"drawing" too, this sets the stage for that and it made it easier to
do the above upload/save tool.