shot tmp.
shot - add shot editor with arbitrary cropping and drawing tools

We don't need padded vs other shot of window now as u can remove the
padding with a single click in the editor now, so remove that (comment
it out).

The editor allows you to click on each screen, on a window (and click
again on a window to remove the padding area for the crop so its
snugly around the window).

You can zoom in and out and scroll around.

You can add lines with an arrow, double-arrowed lines, plain lines
with no arrows, some clipart (e-logo, mallocs, pointing finger... just
having some fun) and some fun "text boxes" like though bubbles,
speecvh bubbles, just plain text with a darking outline, a plain
entry-styled text box and you can select the colors for the text and
background area as well.

It's inksape-like in that the items you draw are objects that can then
be modified or deleted. modifying them raises them and places some
controls on/over/around the object.

And you can still save and share the results.

Fun times!

@feat